webapp/requirements.txt
Jinja2==2.11.2
Werkzeug==1.0.1
uvicorn==0.15.0
fastapi==0.68.1
line-bot-sdk==1.20.0
aiolinebot==0.4 #alpineだと入らない
torch==1.9.1
regex==2021.9.24
mecab-python3==1.0.4
transformers==4.10.3
datasets==1.12.1
fugashi[unidic-lite]==1.1.1
ipadic==1.0.0